spinning-products-for-different-collection
In this, we created metafields from the backend of Shopify to select products for a unique collection. We then display those products in a collection like a spin, allowing the user to rotate through them from the frontend. The selected product appears in the center as a larger image. By clicking on it, the user is redirected to the PDP (Product Detail Page) of that product.