show-menu-item-based-on-url
In this, we added a custom item to the menu that only appears if the user opens the site with a particular link. For example, it will show if the user opens the site with a URL containing "waveLink". Additionally, it stores the user's cookies so that the menu item will remain visible to that user even if they navigate to a different page. The menu item will continue to be shown until the user closes their browser.